refactor(xlsx): ENV-only recalc config, XLSX_RECALC_STRICT, remove financial-modeling mode
- Remove `recalc` and `financial_modeling` tool arguments; recalculation
  behavior is now driven entirely by environment variables.
- Add `XLSX_RECALC_STRICT` env var (default false): when true, formula
  errors and circular references fail the call; when false, errors are
  logged and the file is delivered.
- Delete financial-modeling mode: `apply_financial_styling`, CFA color
  constants, `_promote_variant_for_financial`, `_is_year_string`,
  `FINANCIAL_DEFAULT_*` and all related branches removed from helpers.py.
  General-purpose styling (number-format variants, sources directive,
  default font, formula fill) is preserved.
- Replace test_xlsx_financial.py with test_xlsx_styling.py (general tests
  only); rewrite recalc/edge-case tests to drive behavior via config
  patching; add XLSX_RECALC_STRICT coverage.
- Update AGENTS.md and Readme.md to reflect ENV-only recalc and removal
  of financial-modeling mode.
