+++ main.py
@@ -154,12 +154,10 @@ else:
     annotations={"title": "Markdown to Excel Converter"}
 )
 async def create_excel_document(
-    markdown_content: Annotated[str, Field(description="Markdown content containing tables, headers, and formulas. Use '## Sheet: Sheet Name' to create multiple worksheets. Formulas MUST start with '=' prefix (e.g. =SUM(A1:A5), =A1/B1*100). Use T1.B[0] for cross-table references and B[0] for current row references. Use SheetName!T1.B[0] for cross-sheet references (resolves to SheetName!B2 in Excel). External-workbook references use square brackets: =[Forecast.xlsx]Sheet1!A1. ALWAYS use [0], [1], [2] notation, NEVER use absolute row numbers like B2, B3. Do NOT count table header as first row, first row has index [0]. Supports cell formatting: **bold**, *italic*. Supports column alignment via separator row: |:---|  left, |:---:| center, |---:| right. Dates are auto-detected and formatted (ISO, European, US formats). Plain numeric cells preserve decimals (e.g. 1500.75 stays 1500.75, not rounded to 1,501); values with thousands separators like 1,234 are parsed as numbers automatically. Supports HTML comment directives placed on a line directly above a table: '<!-- freeze -->' freezes the header row so it stays visible when scrolling; '<!-- types: text, currency:$, date, bool, number, percent, multiple -->' forces column data types per column (comma-separated, one per column; omit or leave blank for auto-detection). Type options: 'text' keeps value as-is (preserves leading zeros), 'currency:<symbol>' or 'currency:<symbol>:<variant>' strips symbol and stores as number with currency format (symbols: $, €, £, ¥, Kč, zł, kr, CHF, R$, ₹; use variant 'integer' (or 'int'/'whole') for zero-decimal whole-unit amounts like revenue in $mm — 'currency:$:integer' → $#,##0; combine with dash/parens as 'currency:$:integer:dash'), 'date' or 'date:<format>' parses date with optional Excel format, 'bool' maps true/false/yes/no to Excel boolean, 'number', 'number:<format>', or 'number:<variant>' ensures numeric with optional format or variant, 'percent' or 'percent:<variant>' converts '50%' to 0.5 with percentage format (default 0.0% — one decimal per CFA convention; use 'percent:integer' for 0%), 'multiple' or 'multiple:<variant>' (also accepted as 'number:multiple' alias) formats valuation multiples as '12.5x' (value stored raw, 'x' is display only; trailing 'x' in input is stripped). Variants control negative/zero rendering: 'dash' shows zeros as '-' and negatives in parentheses, 'parens' shows negatives in parentheses. '<!-- sources: B2=Source text, B5=Another -->' attaches cell comments for source citations (supports ranges like B2:B5; if a single-cell entry and a range both cover the same cell, the single-cell entry wins). Directives can be placed directly above a table OR above a '## Sheet:' header that precedes the table — both work, and a directive applies to exactly one table. Recommended source format: name where it came from, when, and a pointer back — e.g. 'Source: Company 10-K, FY2024, Page 45, Revenue Note'. Multiple directives can be stacked above the same table. FINANCIAL MODELING CONVENTIONS (apply these when building financial models, budgets, or forecasts even without financial_modeling=true): (1) Put assumptions (growth rates, margins, multiples) in their own cells and reference them in formulas — write '=B5*(1+$B$6)' not '=B5*1.05' so the model recalculates when the assumption changes; (2) Always state units in column headers — 'Revenue ($mm)', 'Growth Rate (%)', not bare 'Revenue'; (3) Use formulas for every calculation (totals, growth, ratios) rather than hardcoding computed values, so the spreadsheet stays dynamic. NOTE: when financial_modeling=true, numeric and percent cells automatically use the dash number format (zeros render as '-', negatives in parentheses) per CFA convention, and circular-reference detection always runs (even if recalc is disabled) — cycles are logged and, when recalc is explicitly enabled, fail the call.")],
+    markdown_content: Annotated[str, Field(description="Markdown content containing tables, headers, and formulas. Use '## Sheet: Sheet Name' to create multiple worksheets. Formulas MUST start with '=' prefix (e.g. =SUM(A1:A5), =A1/B1*100). Use T1.B[0] for cross-table references and B[0] for current row references. Use SheetName!T1.B[0] for cross-sheet references (resolves to SheetName!B2 in Excel). External-workbook references use square brackets: =[Forecast.xlsx]Sheet1!A1. ALWAYS use [0], [1], [2] notation, NEVER use absolute row numbers like B2, B3. Do NOT count table header as first row, first row has index [0]. Supports cell formatting: **bold**, *italic*. Supports column alignment via separator row: |:---|  left, |:---:| center, |---:| right. Dates are auto-detected and formatted (ISO, European, US formats). Plain numeric cells preserve decimals (e.g. 1500.75 stays 1500.75, not rounded to 1,501); values with thousands separators like 1,234 are parsed as numbers automatically. Supports HTML comment directives placed on a line directly above a table: '<!-- freeze -->' freezes the header row so it stays visible when scrolling; '<!-- types: text, currency:$, date, bool, number, percent, multiple -->' forces column data types per column (comma-separated, one per column; omit or leave blank for auto-detection). Type options: 'text' keeps value as-is (preserves leading zeros), 'currency:<symbol>' or 'currency:<symbol>:<variant>' strips symbol and stores as number with currency format (symbols: $, €, £, ¥, Kč, zł, kr, CHF, R$, ₹; use variant 'integer' (or 'int'/'whole') for zero-decimal whole-unit amounts like revenue in $mm — 'currency:$:integer' → $#,##0; combine with dash/parens as 'currency:$:integer:dash'), 'date' or 'date:<format>' parses date with optional Excel format, 'bool' maps true/false/yes/no to Excel boolean, 'number', 'number:<format>', or 'number:<variant>' ensures numeric with optional format or variant, 'percent' or 'percent:<variant>' converts '50%' to 0.5 with percentage format (default 0.0%; use 'percent:integer' for 0%), 'multiple' or 'multiple:<variant>' (also accepted as 'number:multiple' alias) formats valuation multiples as '12.5x' (value stored raw, 'x' is display only; trailing 'x' in input is stripped). Variants control negative/zero rendering: 'dash' shows zeros as '-' and negatives in parentheses, 'parens' shows negatives in parentheses. '<!-- sources: B2=Source text, B5=Another -->' attaches cell comments for source citations (supports ranges like B2:B5; if a single-cell entry and a range both cover the same cell, the single-cell entry wins). Directives can be placed directly above a table OR above a '## Sheet:' header that precedes the table — both work, and a directive applies to exactly one table. Recommended source format: name where it came from, when, and a pointer back — e.g. 'Source: Company 10-K, FY2024, Page 45, Revenue Note'. Multiple directives can be stacked above the same table. Formula recalculation (cached values for preview) is controlled server-side via XLSX_RECALC_ENABLED (default true); circular-reference detection always runs and reports #CIRC! cycles in server logs.")],
     file_name: Annotated[Optional[str], Field(description="Custom filename for the output file (without extension). If not provided, a unique identifier will be used.", default=None)] = None,
     auto_filter: Annotated[bool, Field(description="If true, adds Excel auto-filter dropdowns to table headers, enabling sorting and filtering in Excel.", default=False)] = False,
     default_font: Annotated[Optional[str], Field(description="Optional font family applied to every cell in the workbook (e.g. 'Arial', 'Calibri', 'Times New Roman'). When omitted, falls back to the server's XLSX_DEFAULT_FONT config value, then openpyxl's default (Calibri).", default=None)] = None,
-    financial_modeling: Annotated[bool, Field(description="When true, applies CFA-standard financial-modeling conventions: (1) color-coding — hardcoded inputs in blue, local formulas in black, cross-sheet formulas in green, external-workbook links in red, source-cited cells get a yellow background; (2) 4-digit-year strings in data rows are treated as text labels rather than numbers; (3) typically combined with the 'sources:' directive and number-format variants like 'number:dash' or 'multiple'. Useful for financial models, budgets, and forecasts.", default=False)] = False,
-    recalc: Annotated[Optional[bool], Field(description="When true, evaluates every formula in-process using a pure-Python engine and writes the computed values back into the file as cached values, so the file previews correctly in tools that don't recalculate on open (Google Sheets preview, mail clients, openpyxl data_only=True). When false, formulas are written without cached values (Excel/LibreOffice will recalc on open). When null (default), falls back to the server's XLSX_RECALC_ENABLED config value (default true). ERROR POLICY: when recalc is explicitly set to true and formula errors are detected (#REF!, #DIV/0!, #VALUE!, #NAME?, #NULL!, #NUM!, #N/A, plus circular references reported as #CIRC!), the tool call FAILS with a descriptive error grouping errors by type with counts and locations, so the model can fix all errors of a given kind at once — enforcing a zero-errors delivery standard. Circular references are detected via independent graph analysis and are reported even when the recalc engine is unavailable. When recalc runs as a defaulted behaviour (parameter not passed), errors are logged but the file is still delivered. Recalculation is bounded by XLSX_RECALC_TIMEOUT_SECONDS (default 30s); on timeout the file is delivered without cached values. CACHED VALUE SUPPORT: numeric, boolean, datetime, AND string results are all cached (string formulas like =A1&' total' or =IF(A1>0,'Yes','No') use OOXML inline-string t='str' cells, so they preview correctly in non-recalculating clients).", default=None)] = None,
 ) -> str:
     """
     Converts markdown to Excel with advanced formula support.
@@ -174,8 +172,6 @@ async def create_excel_document(
             file_name=file_name,
             auto_filter=auto_filter,
             default_font=default_font,
-            financial_modeling=financial_modeling,
-            recalc=recalc,
         )
         logger.info("Excel document uploaded successfully")
         return result
